chore: moved giving and commons to the folder
