+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/account/adyenAccount.js
@@ -1,6 +1,6 @@
 const store = require('../../../../../config/store');
-const { onFieldValid, onBrand, getPaymentMethods } = require('../../commons');
-const { httpClient } = require('../../commons/httpClient');
+const { onFieldValid, onBrand, getPaymentMethods } = require('../commons');
+const { httpClient } = require('../commons/httpClient');
 
 let checkout;
 let card;
