+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/__tests__/renderGiftcardComponent.test.js
@@ -9,7 +9,7 @@ const {
 } = require('../giftcards');
 const store = require('../../../../../../config/store');
 
-jest.mock('../../../commons');
+jest.mock('../../commons');
 jest.mock('../../../../../../config/store');
 
 beforeEach(() => {
