+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/amazonPayCheckout.js
@@ -1,7 +1,7 @@
 const store = require('../../../../../config/store');
 const helpers = require('./helpers');
 const { AMAZON_PAY } = require('../../../../../config/constants');
-const { httpClient } = require('../../commons/httpClient');
+const { httpClient } = require('../commons/httpClient');
 
 function handleAuthorised(response) {
   document.querySelector('#result').value = JSON.stringify({
