+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/checkout/paymentMethodsConfiguration/index.js
@@ -12,7 +12,7 @@ const PayPalConfig = require('./paypal/paypalConfig');
 const GiftCardsConfig = require('./giftcards/giftcardsConfig');
 const store = require('../../../../../../config/store');
 const helpers = require('../helpers');
-const { httpClient } = require('../../../commons/httpClient');
+const { httpClient } = require('../../commons/httpClient');
 
 function getPaymentMethodsConfiguration(email, paymentMethodsResponse) {
   const { amount, AdyenPaymentMethods } = paymentMethodsResponse;
