+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/cart/expressPayments.js
@@ -1,4 +1,4 @@
-const { getPaymentMethods } = require('../../../commons');
+const { getPaymentMethods } = require('../../commons');
 const {
   Paypal,
   ApplePay,
