+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/paymentMethods/amazonpay/amazonPayExpressPart2.js
@@ -1,6 +1,6 @@
-const { getPaymentMethods } = require('../../../../commons');
+const { getPaymentMethods } = require('../../../commons');
 const { AMAZON_PAY } = require('../../../../../../../config/constants');
-const { httpClient } = require('../../../../commons/httpClient');
+const { httpClient } = require('../../../commons/httpClient');
 
 async function saveShopperDetails(details) {
   const data = await httpClient({
