+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/paymentMethods/googlepay/googlepay.js
@@ -1,4 +1,4 @@
-const { httpClient } = require('../../../../commons/httpClient');
+const { httpClient } = require('../../../commons/httpClient');
 const store = require('../../../../../../../config/store');
 const helpers = require('../../../checkout/helpers');
 const {
@@ -6,7 +6,7 @@ const {
   GOOGLE_PAY_CALLBACK_TRIGGERS,
 } = require('../../../../../../../config/constants');
 const { initializeCheckout } = require('../../initializeCheckout');
-const { createTemporaryBasket } = require('../../../../commons');
+const { createTemporaryBasket } = require('../../../commons');
 
 class GooglePay {
   constructor(config, applicationInfo, adyenTranslations, isExpressPdp) {
