+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen/express/product/expressPayments.js
@@ -3,8 +3,8 @@ const {
   GOOGLE_PAY,
   PAY_WITH_GOOGLE,
 } = require('../../../../../../config/constants');
-const { getPaymentMethods } = require('../../../commons/index');
-const { httpClient } = require('../../../commons/httpClient');
+const { getPaymentMethods } = require('../../commons/index');
+const { httpClient } = require('../../commons/httpClient');
 const { ApplePay, GooglePay } = require('../paymentMethods');
 
 function getProductForm(product) {
